Address page (#448)
* Abstract SubDomains component into standalone component

* Fix import

* Basic address page with abstracted components

* Add ExternalLink to etherscan

* Adjust margin

* Edit contenthash placeholder to add onion3

* fix camelcase

* Fix typo

* Remove errors

* Fix dnssec bug

* Remove log
