+++ src/components/Address/index.js
@@ -1,0 +1,2 @@
+import Address from './Address'
+export default Address
