+++ src/components/Basic/TopBar.js
@@ -1,0 +1,18 @@
+import styled from '@emotion/styled'
+
+const TopBar = styled('div')`
+  padding: 20px 40px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  border-bottom: 1px solid #ededed;
+  box-shadow: 0 2px 4px 0 rgba(181, 177, 177, 0.2);
+
+  background: ${({ percentDone }) =>
+    percentDone
+      ? `
+  linear-gradient(to right, rgba(128, 255, 128, 0.1) 0%, rgba(82,229,255, 0.1) ${percentDone}%,#ffffff ${percentDone}%)`
+      : 'white'};
+`
+
+export default TopBar
