+++ src/components/SingleName/Name.js
@@ -1,10 +1,11 @@
 import React from 'react'
 import styled from '@emotion/styled'
 
-import mq, { useMediaMin } from 'mediaQuery'
+import { useMediaMin } from 'mediaQuery'
 import { EMPTY_ADDRESS } from '../../utils/records'
 
 import { Title } from '../Typography/Basic'
+import TopBar from '../Basic/TopBar'
 import DefaultFavourite from '../AddFavourite/Favourite'
 import NameDetails from './NameDetails'
 import NameRegister from './NameRegister'
@@ -12,59 +13,7 @@ import DNSNameRegister from './DNSNameRegister'
 import ShortName from './ShortName'
 import Tabs from './Tabs'
 import { useAccount } from '../QueryAccount'
-
-const NameContainer = styled('div')`
-  background: white;
-  box-shadow: 3px 4px 6px 0 rgba(229, 236, 241, 0.3);
-  border-radius: 0;
-  margin-bottom: 60px;
-  position: relative;
-  overflow: hidden;
-
-  ${mq.small`
-    border-radius: 6px;
-  `}
-
-  &:before {
-    left: 0;
-    top: 0;
-    width: 4px;
-    height: 100%;
-    display: block;
-    content: '';
-    background: ${({ state }) => {
-      switch (state) {
-        case 'Owned':
-          return '#CACACA'
-        case 'Auction':
-        case 'Reveal':
-          return 'linear-gradient(-180deg, #42E068 0%, #52E5FF 100%)'
-        case 'Yours':
-          return '#52e5ff'
-        case 'Open':
-          return '#42E068'
-        default:
-          return '#CACACA'
-      }
-    }};
-    position: absolute;
-  }
-`
-
-const TopBar = styled('div')`
-  padding: 20px 40px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  border-bottom: 1px solid #ededed;
-  box-shadow: 0 2px 4px 0 rgba(181, 177, 177, 0.2);
-
-  background: ${({ percentDone }) =>
-    percentDone
-      ? `
-  linear-gradient(to right, rgba(128, 255, 128, 0.1) 0%, rgba(82,229,255, 0.1) ${percentDone}%,#ffffff ${percentDone}%)`
-      : 'white'};
-`
+import NameContainer from '../Basic/MainContainer'
 
 const Owner = styled('div')`
   color: #ccd4da;
