+++ src/components/SingleName/SubDomains.js
@@ -1,5 +1,4 @@
 import React from 'react'
-import { Link } from 'react-router-dom'
 import styled from '@emotion/styled'
 import { Query } from 'react-apollo'
 import {
@@ -8,8 +7,8 @@ import {
 } from '../../graphql/queries'
 import Loader from '../Loader'
 import { H2 } from '../Typography/Basic'
-import { SingleNameBlockies } from './SingleNameBlockies'
 import AddSubdomain from './AddSubdomain'
+import ChildDomainItem from '../DomainItem/ChildDomainItem'
 import { getNamehash, encodeLabelhash } from '@ensdomains/ui'
 
 const SubDomainsContainer = styled('div')`
@@ -22,21 +21,6 @@ const SubDomainH2 = styled(H2)`
   padding: 20px 0 50px;
   text-align: center;
   color: #ccd4da;
-`
-
-const SubDomainLink = styled(Link)`
-  display: flex;
-  align-items: center;
-  width: 100%;
-  padding: 30px 0;
-  color: #2b2b2b;
-  font-size: 22px;
-  font-weight: 100;
-  border-bottom: 1px dashed #d3d3d3;
-
-  &:last-child {
-    border: none;
-  }
 `
 
 function SubDomainsFromWeb3(domain, isOwner) {
@@ -66,12 +50,12 @@ function SubDomainsFromWeb3(domain, isOwner) {
             {isOwner && <AddSubdomain domain={domain} refetch={refetch} />}
             {data &&
               data.getSubDomains.subDomains.map(d => (
-                <SubDomainLink key={d.name} to={`/name/${d.name}`}>
-                  <SingleNameBlockies imageSize={24} address={d.owner} />
-                  {d.decrypted
-                    ? d.name
-                    : `[unknown${d.labelHash.slice(2, 12)}].${d.node}`}
-                </SubDomainLink>
+                <ChildDomainItem
+                  name={d.name}
+                  owner={d.owner}
+                  parent={d.node}
+                  labelhash={d.labelHash}
+                />
               ))}
           </>
         )
@@ -132,17 +116,12 @@ function SubDomains({ domain, isOwner, ...rest }) {
                       name = `${encodeLabelhash(d.labelhash)}.${domain.name}`
                     }
                     return (
-                      <SubDomainLink key={name} to={`/name/${name}`}>
-                        <SingleNameBlockies
-                          imageSize={24}
-                          address={d.owner.id}
-                        />
-                        {d.labelName !== null
-                          ? `${name}`
-                          : `[unknown${d.labelhash.slice(2, 10)}].${
-                              domain.name
-                            }`}
-                      </SubDomainLink>
+                      <ChildDomainItem
+                        name={name}
+                        owner={d.owner.id}
+                        parent={domain.name}
+                        labelhash={d.labelHash}
+                      />
                     )
                   })}
               </>
