+++ src/routes/AddressPage.js
@@ -1,0 +1,9 @@
+import React from 'react'
+import Address from '../components/Address'
+
+const AddressContainer = ({ match }) => {
+  const address = match.params.address.toLowerCase()
+  return <Address address={address} />
+}
+
+export default AddressContainer
